feat: updated frontend for casing deletion process fix
